+++ app/Product.php
@@ -76,7 +76,7 @@ class Product extends Model
 
     public function setWeightAttribute($value)
     {
-        $this->attributes['weight'] = (int)$value;
+        $this->attributes['weight'] = (float)$value;
     }
